Update bot.py
Added in a "meme" function, will pull images from a directory. Must be a direct path.
Added in a kill bot function, will terminate the session - any role can do this, for now.
Note: Time function is set to whatever computer the bot runs in, change Australia to whatever country you're in.